+++ action_server/src/sema4ai/action_server/_robo_utils/process.py
@@ -10,6 +10,9 @@ import threading
 from pathlib import Path
 from typing import Dict, List, Optional, Protocol, Union
 
+# Note: keep this import for backward compatibility
+from sema4ai.common.process import kill_process_and_subprocesses
+
 from sema4ai.action_server._preload_actions.preload_actions_autoexit import (
     is_process_alive,
 )
@@ -257,72 +260,6 @@ def _call(cmdline, **kwargs):
     except Exception:
         log.exception("Error running: %s", (" ".join(cmdline)))
         return None
-
-
-def _kill_process_and_subprocess_linux(pid):
-    initial_pid = pid
-
-    def list_children_and_stop_forking(ppid):
-        children_pids = []
-        _call(
-            ["kill", "-STOP", str(ppid)], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-
-        list_popen = _popen(
-            ["pgrep", "-P", str(ppid)], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-
-        if list_popen is not None:
-            stdout, _ = list_popen.communicate()
-            for line in stdout.splitlines():
-                line = line.decode("ascii").strip()
-                if line:
-                    pid = str(line)
-                    children_pids.append(pid)
-                    # Recursively get children.
-                    children_pids.extend(list_children_and_stop_forking(pid))
-        return children_pids
-
-    previously_found = set()
-
-    for _ in range(50):  # Try this at most 50 times before giving up.
-        children_pids = list_children_and_stop_forking(initial_pid)
-        found_new = False
-
-        for pid in children_pids:
-            if pid not in previously_found:
-                found_new = True
-                previously_found.add(pid)
-                _call(
-                    ["kill", "-KILL", str(pid)],
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.PIPE,
-                )
-
-        if not found_new:
-            break
-
-    # Now, finish the initial one.
-    _call(
-        ["kill", "-KILL", str(initial_pid)],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    )
-
-
-def kill_process_and_subprocesses(pid):
-    log.debug("Killing process and subprocesses of: %s", pid)
-    from subprocess import CalledProcessError
-
-    if sys.platform == "win32":
-        args = ["taskkill", "/F", "/PID", str(pid), "/T"]
-        retcode = subprocess.call(
-            args, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE
-        )
-        if retcode not in (0, 128, 255):
-            raise CalledProcessError(retcode, args)
-    else:
-        _kill_process_and_subprocess_linux(pid)
 
 
 class IProgressReporter(Protocol):