+++ action_server/tests/conftest.py
@@ -3,3 +3,9 @@ pytest_plugins = [
     "action_server_tests.fixtures",
     "action_server_tests.data_server_cli_fixtures",
 ]
+
+
+def pytest_configure(config):
+    config.addinivalue_line(
+        "markers", "integration_test: mark test to run only on integration test"
+    )